+++ base/base.vcxproj.filters
@@ -35,10 +35,16 @@
     <ClInclude Include="fingerprint2011.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="monostable.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="not_null_test.cpp">
       <Filter>Test Files</Filter>
+    </ClCompile>
+    <ClCompile Include="monostable.cpp">
+      <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
