+++ base/monostable.cpp
@@ -1,7 +1,7 @@
-#include "ksp_plugin/monostable.hpp"
+#include "base/monostable.hpp"
 
 namespace principia {
-namespace ksp_plugin {
+namespace base {
 
 void Monostable::Flop() {
   transient_ = false;
@@ -11,5 +11,5 @@ Monostable::operator bool() const {
   return transient_;
 }
 
-}  // namespace ksp_plugin
+}  // namespace base
 }  // namespace principia
