+++ ksp_plugin/ksp_plugin.vcxproj
@@ -133,7 +133,6 @@
     <ClInclude Include="celestial_body.hpp" />
     <ClInclude Include="frames.hpp" />
     <ClInclude Include="mock_plugin.hpp" />
-    <ClInclude Include="monostable.hpp" />
     <ClInclude Include="part.hpp" />
     <ClInclude Include="part_body.hpp" />
     <ClInclude Include="physics_bubble.hpp" />
@@ -145,7 +144,6 @@
   <ItemGroup>
     <ClCompile Include="interface.cpp" />
     <ClCompile Include="mock_plugin.cpp" />
-    <ClCompile Include="monostable.cpp" />
     <ClCompile Include="physics_bubble.cpp" />
     <ClCompile Include="plugin.cpp" />
   </ItemGroup>
