+++ ksp_plugin/ksp_plugin.vcxproj.filters
@@ -20,9 +20,6 @@
     <ClInclude Include="mock_plugin.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="monostable.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="vessel.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -53,9 +50,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="plugin.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="monostable.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="mock_plugin.cpp">
